Unfork courseware test_module_render
